% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dams-package.R
\docType{package}
\name{dams-package}
\alias{dams-package}
\title{Dams in the United States from the National Inventory of Dams (NID)}
\description{
Data from NID was downloaded from \url{http://nid.usace.army.mil}.
Subsequently, the raw data was checked for potential errors and cleaned.
The dams package provides a subset of NID fields and functionality (\code{\link[=get_nid]{get_nid()}}) to access the entire NID dataset.
}
