% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{package}
\name{dams}
\alias{dams}
\alias{dams-package}
\title{Dams in the United States from the National Inventory of Dams (NID).}
\description{
Data from NID was extracted manually from <http://nid.usace.army.mil>. 
Subsequently, the raw data was checked for potential errors and cleaned. 
dams package provides sample cleaned data from the NID and provides 
functionality to access the entire cleaned NID data.
}

