% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dams-package.R
\docType{data}
\name{nid_subset}
\alias{nid_subset}
\alias{nid_cleaned}
\title{Subset of dams information from the NID database}
\format{
Data frame with 32 columns and 91457 rows
}
\usage{
data(nid_subset)
}
\description{
Reduced subset of the the NID data excluding fields with more than 5 percent missing data. See the \link{nid_all} documentation for a description of each field.
}
\seealso{
\link{nid_all} for a description of each field.
}
\keyword{datasets}
