% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nid.R
\name{get_nid}
\alias{get_nid}
\title{Retrieve \link{nid_all} from the official NID site}
\usage{
get_nid(dest = "NID2019_U.xlsx", overwrite = FALSE)
}
\arguments{
\item{dest}{destination file path}

\item{overwrite}{logical. overwrite.}
}
\value{
\link{nid_all} entire NID data, all the 74000+ records from \url{http://nid.usace.army.mil/}
}
\description{
Retrieve \link{nid_all} from the official NID site
}
\examples{
\dontrun{
dams_all <- get_nid()
}

}
