% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nid.r
\name{get_nid}
\alias{get_nid}
\title{Retrieve desired data on dams from pre-processed NID data}
\usage{
get_nid()
}
\description{
Retrieve desired data on dams from pre-processed NID data
}
\examples{

# entire NID data, all the 74000+ records from bitbucket.org/rationshop
\dontrun{
dams_all <- get_nid()
}

}

